requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
alembic==1.4.3
    # via flask-migrate
apscheduler==3.7.0
    # via flask-apscheduler
beautifulsoup4==4.9.3
    # via bs4
bs4==0.0.1
    # via -r requirements.in
camelot-py[cv]==0.9.0
    # via -r requirements.in
certifi==2020.12.5
    # via requests
cffi==1.14.6
    # via cryptography
chardet==4.0.0
    # via
    #   camelot-py
    #   pdfminer.six
    #   requests
click==7.1.2
    # via
    #   camelot-py
    #   flask
cryptography==3.4.7
    # via pdfminer.six
cycler==0.10.0
    # via matplotlib
et-xmlfile==1.1.0
    # via openpyxl
flask-apscheduler==1.12.1
    # via -r requirements.in
flask-migrate==2.5.3
    # via -r requirements.in
flask-script==2.0.6
    # via -r requirements.in
flask-sqlalchemy==2.4.4
    # via
    #   -r requirements.in
    #   flask-migrate
flask-wtf==0.14.3
    # via -r requirements.in
flask==1.1.2
    # via
    #   -r requirements.in
    #   flask-apscheduler
    #   flask-migrate
    #   flask-script
    #   flask-sqlalchemy
    #   flask-wtf
gitdb==4.0.5
    # via gitpython
gitpython==3.1.8
    # via -r requirements.in
gunicorn==20.0.4
    # via -r requirements.in
idna==2.10
    # via requests
is-safe-url==1.0
    # via -r requirements.in
itsdangerous==1.1.0
    # via
    #   flask
    #   flask-wtf
jinja2==2.11.2
    # via flask
kiwisolver==1.3.1
    # via matplotlib
lxml==4.6.3
    # via -r requirements.in
mako==1.1.3
    # via alembic
markupsafe==1.1.1
    # via
    #   jinja2
    #   mako
    #   wtforms
matplotlib==3.3.3
    # via
    #   -r requirements.in
    #   mplfinance
mplfinance==0.12.7a10
    # via -r requirements.in
numpy==1.19.5
    # via
    #   -r requirements.in
    #   camelot-py
    #   matplotlib
    #   opencv-python
    #   pandas
opencv-python==4.5.2.54
    # via camelot-py
openpyxl==3.0.7
    # via camelot-py
pandas==1.2.3
    # via
    #   -r requirements.in
    #   camelot-py
    #   mplfinance
pdfminer.six==20201018
    # via camelot-py
pillow==8.1.0
    # via matplotlib
psycopg2-binary==2.8.6
    # via -r requirements.in
pycparser==2.20
    # via cffi
pyparsing==2.4.7
    # via matplotlib
pypdf2==1.26.0
    # via camelot-py
python-dateutil==2.8.1
    # via
    #   alembic
    #   flask-apscheduler
    #   matplotlib
    #   pandas
python-dotenv==0.15.0
    # via -r requirements.in
python-editor==1.0.4
    # via alembic
pytz==2021.1
    # via
    #   apscheduler
    #   pandas
    #   tzlocal
requests==2.25.1
    # via -r requirements.in
six==1.15.0
    # via
    #   apscheduler
    #   cycler
    #   python-dateutil
smmap==3.0.4
    # via gitdb
sortedcontainers==2.4.0
    # via pdfminer.six
soupsieve==2.2.1
    # via beautifulsoup4
sqlalchemy==1.3.19
    # via
    #   alembic
    #   flask-sqlalchemy
tzlocal==2.1
    # via apscheduler
urllib3==1.26.4
    # via requests
werkzeug==1.0.1
    # via flask
wtforms==2.3.3
    # via flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
